Move shared navigation specific files to src/components/navigation
