+++ app/packs/src/components/navigation/NavNewSession.js
